% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_functions.R
\name{genhaps}
\alias{genhaps}
\title{Generate haplotypes}
\usage{
genhaps(data, slist = colnames(data), snps, samples, A1 = "allele.A",
  A2 = "allele.B", cols.copy = character(0), f.in = "snphap/haps.in",
  redo = FALSE)
}
\arguments{
\item{data}{snpStats object}

\item{slist}{snp list}

\item{snps}{snps dataframe with columns "allele.A", "allele.B" denoting alleles}

\item{samples}{samples dataframe}

\item{A1}{colname of allele 1 in snps}

\item{A2}{colname of allele 2 in snps}

\item{cols.copy}{columns to be copied to resulting phased object}

\item{f.in}{filename for snphap input (will be overwritten)}

\item{redo}{set to TRUE to regenerate snphap output regardless of whether it exists}
}
\value{
data.frame containing the best phased haplotypes
}
\description{
Write a snpStats object to file, and use snphap to phase with multiple imputation
}
\author{
Chris Wallace
}
