% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_functions.R
\name{snphap.exec}
\alias{snphap.exec}
\title{.. content for \description{} (no empty lines) ..}
\usage{
snphap.exec()
}
\value{
string with location of snphap executable
}
\description{
.. content for \details{} ..
}
\author{
Chris Wallaced
}
